orvim-models/clip/requirements.txt
fastapi
uvicorn
transformers
pillow
torch